requirements.txt
# 核心依赖
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0

# CAD文件支持 (可选)
ezdxf>=0.18.0

# PDF导出支持 (可选)
reportlab>=3.6.0

# 图像处理
opencv-python>=4.5.0

# 数学和科学计算
scipy>=1.7.0

# 数据处理
pandas>=1.3.0

# GUI增强
ttkthemes>=3.2.0

# 配置文件处理
pyyaml>=6.0

# 日志记录
loguru>=0.6.0

# 进度条
tqdm>=4.62.0

# 并行处理
joblib>=1.1.0